Auth0用に最適化
