+++ src/api/authApi.ts
@@ -1,6 +1,5 @@
-import { axiosAuthClient } from "./axiosAuthClient";
+import { axiosClient } from "./axiosClient";
 
 export const authApi = {
-  register: (params: string) => axiosAuthClient.post("/register", params),
-  login: (params: string) => axiosAuthClient.post("/login", params),
+  login: () => axiosClient.post("/login"),
 };
