+++ src/api/axiosClient.ts
@@ -1,6 +1,4 @@
 import axios from "axios";
-import { useRecoilValue } from "recoil";
-import { tokenAtom } from "../recoil/TokenAtom";
 
 export const axiosClient = axios.create({
   baseURL: process.env.REACT_APP_API_URL,
@@ -10,8 +8,6 @@ const getToken = () => localStorage.getItem("token");
 
 // APIを叩く前に前処理を行う
 axiosClient.interceptors.request.use(async (config: any) => {
-  // const token = useRecoilValue(tokenAtom);
-  // return config;
   return {
     ...config,
     headers: {
